[#17] - move Alert component to support folder
