+++ src/components/QuestionDefinition.jsx
@@ -5,9 +5,8 @@ import { UIStore, questionFn, questionGroupFn } from '../lib/store';
 import data from '../lib/data';
 import QuestionSetting from './QuestionSetting';
 import QuestionSkipLogic from './QuestionSkipLogic';
-import { AddMoveButton, CardTitle } from '../support';
+import { AddMoveButton, Alert, CardTitle } from '../support';
 import { orderBy, maxBy, minBy } from 'lodash';
-import Alert from './Alert';
 
 const QuestionDefinition = ({ index, question, questionGroup, isLastItem }) => {
   const { questionGroups } = questionGroupFn.store.useState((s) => s);
