+++ src/components/QuestionGroupDefinition.jsx
@@ -3,9 +3,8 @@ import { Card } from 'antd';
 import { UIStore, questionGroupFn } from '../lib/store';
 import QuestionGroupSetting from './QuestionGroupSetting';
 import QuestionDefinition from './QuestionDefinition';
-import { AddMoveButton, CardTitle } from '../support';
+import { AddMoveButton, Alert, CardTitle } from '../support';
 import { orderBy, maxBy, minBy } from 'lodash';
-import Alert from './Alert';
 
 const QuestionGroupDefinition = ({ index, questionGroup, isLastItem }) => {
   const { questionGroups } = questionGroupFn.store.useState((s) => s);
