+++ src/support/index.js
@@ -3,3 +3,4 @@ export { default as CardTitle } from './CardTitle';
 export { default as CardExtraButton } from './CardExtraButton';
 export { default as SaveButton } from './SaveButton';
 export { default as TranslationFormItem } from './TranslationFormItem';
+export { default as Alert } from './Alert';
